requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-multipart>=0.0.9
openpyxl>=3.1.0
pandas>=2.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
slowapi>=0.1.9